Basic mouse movement, added extra tests

+++ STM32/universal_gesture_interface/Core/Inc/motion_processing.h
@@ -10,7 +10,7 @@
 #ifndef INC_MOTION_PROCESSING_H_
 #define INC_MOTION_PROCESSING_H_
 
-void HPF(int16_t* accel, int16_t* prev, int16_t* hp, float alpha);
-void AccelToVel(int16_t* accel, int16_t* vel, float drag, float dt);
+void HPF(int16_t* accel, int16_t* prev, float* hp, float alpha);
+void AccelToVel(float* accel, int16_t* vel, float drag, float dt);
 
 #endif /* INC_MOTION_PROCESSING_H_ */
